+++ src/main/java/com/almostreliable/summoningrituals/compat/kubejs/recipe/component/EntityOutputComponent.java
@@ -3,7 +3,8 @@ package com.almostreliable.summoningrituals.compat.kubejs.recipe.component;
 import com.almostreliable.summoningrituals.SummoningRituals;
 import com.almostreliable.summoningrituals.compat.kubejs.binding.SummoningEntityBinding;
 import com.almostreliable.summoningrituals.compat.kubejs.builder.SummoningEntityOutputBuilder;
-import com.almostreliable.summoningrituals.compat.kubejs.wrapper.SizedEntityTypeWrapper;
+import com.almostreliable.summoningrituals.compat.kubejs.wrapper.EntityComponentTypeWrapper;
+import com.almostreliable.summoningrituals.compat.kubejs.wrapper.EntityInfoTypeWrapper;
 import com.almostreliable.summoningrituals.recipe.EntityInfo;
 import com.almostreliable.summoningrituals.recipe.output.EntityOutput;
 
@@ -35,7 +36,7 @@ public record EntityOutputComponent(RecipeComponentType<?> type) implements Reci
         return TypeInfo.of(EntityOutput.class)
             .or(TypeInfo.of(EntityInfo.class))
             .or(TypeInfo.of(SummoningEntityOutputBuilder.class))
-            .or(TypeInfo.of(Holder.class).withParams(SizedEntityTypeWrapper.ENTITY_TYPE_INFO))
+            .or(TypeInfo.of(Holder.class).withParams(EntityInfoTypeWrapper.ENTITY_TYPE_INFO))
             .or(TypeInfo.of(ResourceLocation.class))
             .or(TypeInfo.STRING);
     }
@@ -50,12 +51,12 @@ public record EntityOutputComponent(RecipeComponentType<?> type) implements Reci
             return new EntityOutput(info, Optional.empty(), Optional.empty());
         }
 
-        return SizedEntityTypeWrapper.wrap(
+        return EntityComponentTypeWrapper.wrap(
             this,
             cx,
             from,
             EntityOutput.class,
-            (entity, count) -> SummoningEntityBinding.output(entity, count).buildOutput()
+            entity -> SummoningEntityBinding.output(entity).buildOutput()
         );
     }
 }