+++ src/main/java/com/almostreliable/summoningrituals/compat/kubejs/wrapper/EntityInfoTypeWrapper.java
@@ -1,34 +1,33 @@
 package com.almostreliable.summoningrituals.compat.kubejs.wrapper;
 
+import com.almostreliable.summoningrituals.compat.kubejs.builder.SummoningEntityBuilder;
+import com.almostreliable.summoningrituals.recipe.EntityInfo;
+
 import net.minecraft.core.Holder;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.neoforged.neoforge.registries.DeferredHolder;
 
-import dev.latvian.mods.kubejs.error.InvalidRecipeComponentValueException;
+import dev.latvian.mods.kubejs.error.KubeRuntimeException;
 import dev.latvian.mods.kubejs.holder.HolderWrapper;
-import dev.latvian.mods.kubejs.recipe.RecipeScriptContext;
-import dev.latvian.mods.kubejs.recipe.component.RecipeComponent;
 import dev.latvian.mods.kubejs.script.KubeJSContext;
+import dev.latvian.mods.kubejs.script.SourceLine;
 import dev.latvian.mods.kubejs.util.Cast;
+import dev.latvian.mods.rhino.Context;
 import dev.latvian.mods.rhino.type.TypeInfo;
+import dev.latvian.mods.rhino.util.wrap.TypeWrapperFactory;
 
-import java.util.function.BiFunction;
+public class EntityInfoTypeWrapper implements TypeWrapperFactory<EntityInfo> {
 
-public final class SizedEntityTypeWrapper {
-
+    public static final EntityInfoTypeWrapper INSTANCE = new EntityInfoTypeWrapper();
+    public static final TypeInfo TYPE_INFO = TypeInfo.of(EntityInfo.class);
     public static final TypeInfo ENTITY_TYPE_INFO = TypeInfo.of(EntityType.class);
 
-    private SizedEntityTypeWrapper() {}
-
-    public static <T> T wrap(
-        RecipeComponent<?> component, RecipeScriptContext cx, Object from, Class<T> type,
-        BiFunction<Holder<EntityType<?>>, Integer, T> factory
-    ) {
+    @Override
+    public EntityInfo wrap(Context cx, Object from, TypeInfo target) {
         var o = from;
-
-        if (type.isInstance(o)) {
-            return type.cast(o);
+        if (o instanceof EntityInfo entityInfo) {
+            return entityInfo;
         }
 
         String sequence = null;
@@ -51,18 +50,15 @@ public final class SizedEntityTypeWrapper {
 
         Holder<EntityType<?>> entityHolder;
         try {
-            entityHolder = Cast.to(HolderWrapper.wrap((KubeJSContext) cx.cx(), o, ENTITY_TYPE_INFO));
+            entityHolder = Cast.to(HolderWrapper.wrap((KubeJSContext) cx, o, ENTITY_TYPE_INFO));
         } catch (Exception e) {
-            throw new InvalidRecipeComponentValueException("invalid summoning entity", component, from).source(cx.recipe().sourceLine);
+            throw new KubeRuntimeException("invalid summoning entity: " + o).source(SourceLine.of(cx));
         }
 
         if (entityHolder instanceof DeferredHolder<?, ?>) {
-            throw new InvalidRecipeComponentValueException(
-                "unknown entity id for summoning entity",
-                component,
-                from
-            ).source(cx.recipe().sourceLine);
+            throw new KubeRuntimeException("unknown entity id for summoning entity: " + o).source(SourceLine.of(cx));
         }
-        return factory.apply(entityHolder, count);
+
+        return new SummoningEntityBuilder(entityHolder, count).build();
     }
 }